`run_batch_script.yml`: Fix `grep` checks by adding `!`
`grep` returns 0 if found, and 1 if not found. But we want the opposite: 1 if warning/error
is present, and 0 if there are no warnings/errors. So, we need to invert the output code.

I have no idea why these checks lasted as long as they did without this `!` inversion. The
`grep` commands only used options `-iF`, meaning (case-insensitive) and (non-regex
strings), so the options don't have anything to do with it.

Hopefully this will fix the erroneous failures?